remove ode_data
