+++ src/jaxsim/math/__init__.py
@@ -1,6 +1,3 @@
-# Define the default standard gravity constant.
-StandardGravity = 9.81
-
 from .adjoint import Adjoint
 from .cross import Cross
 from .inertia import Inertia
@@ -11,3 +8,7 @@ from .transform import Transform
 from .utils import safe_norm
 
 from .joint_model import JointModel, supported_joint_motion  # isort:skip
+
+
+# Define the default standard gravity constant.
+STANDARD_GRAVITY = -9.81
