+++ src/jaxsim/rbda/contacts/relaxed_rigid.py
@@ -337,7 +337,7 @@ class RelaxedRigidContacts(common.ContactModel):
             model=model,
             position_constraint=position_constraint,
             velocity_constraint=velocity,
-            parameters=data.contacts_params,
+            parameters=model.contacts_params,
         )
 
         # Compute the Delassus matrix and the free mixed linear acceleration of
